Move the config string to RunMojo
